+++ REvoDesign/external_designer/__init__.py
@@ -1,0 +1,4 @@
+# ALL External designners must get registed at here.
+from REvoDesign.external_designer.ExternalDesigners import *
+
+EXTERNAL_DESIGNERS = {'ProteinMPNN': ColabDesigner_MPNN}